ref qa

+++ resources/views/components/action-buttons.blade.php
@@ -4,9 +4,11 @@
     <a href="{{ $edit }}"><i class="fas fa-edit"></i></a>
     @endrole
     @if(\Route::current()->getName() == 'users.index')
-        {{ Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $id]]) }}
-        {{ Form::button('<i class="fas fa-trash" aria-hidden="true"></i>',
-            ['class' => 'btn btn-warning btn-sm', 'type' => 'submit', 'style' => 'padding:0; background: none; border:none; color:#007bff;'] )  }}
-        {{ Form::close() }}
+        @if(auth()->id() != $id)
+            {{ Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $id]]) }}
+            {{ Form::button('<i class="fas fa-trash" aria-hidden="true"></i>',
+                ['class' => 'btn btn-warning btn-sm', 'type' => 'submit', 'style' => 'padding:0; background: none; border:none; color:#007bff;'] )  }}
+            {{ Form::close() }}
+        @endif
     @endif
 </td>
